+++ pyffi/spells/nif/optimize.py
@@ -460,8 +460,10 @@ class SpellOptimizeGeometry(pyffi.spells.nif.NifSpell):
                      morph.vectors.update_size()
 
         # recalculate tangent space (only if the branch already exists)
-        if branch.find(block_name='Tangent space (binormal & tangent vectors)',
-                       block_type=NifFormat.NiBinaryExtraData):
+        if (branch.find(block_name='Tangent space (binormal & tangent vectors)',
+                        block_type=NifFormat.NiBinaryExtraData)
+            or (data.num_uv_sets & 61440)
+            or (data.bs_num_uv_sets & 61440)):
             self.toaster.msg("recalculating tangent space")
             branch.update_tangent_space()